requirements.txt
click~=8.1.0
pillow~=11.0.0
playwright~=1.49.0
pypdf~=5.1.0
pyyaml~=6.0.2
